Created an initial Amazon API library.
Made each html object able to load its own css and js.

Added error page templates at both the module level and the framework level.

Added logging of errors and exceptions to a database table.

Added page to view logged errors.

Created a DataTable object.

Created the MarkupTextarea class that allows the management of rich markup.

Added support for Closure JS minification.

Added support for UglifyJS JS minification.

Moved the base configuration file to the protected/ folder.

Added functionality to encrypt the base configuration file upon creation.

Moved the framework and module configurations to the database.

Changed configuration variables to have data types used to keep data integrity.

Added a flag to the form fields that keeps track of if the form was submitted.

Restructured the directory structure so that most PHP code falls behind the web root directory.

Reorganized the directory structure.

Rewrote the autoloader to utilize PHP's built-in spl_autoload with namespaces.

Changed the name of each class file to be lowercase due to unresolved PHP bug #49625 where spl_autoload looks for lowercased file names causing errors on case-sensitive file systems.

Added namespaces to all classes.

Redesigned parts of the admin module's layout.

Rewrote the framework core to be more object-oriented with each mode being a child object of the core.

Rewrote the error handling to be more object-oriented with each framework mode being a child object of the base error class.

Added a backtrace to PHP errors.

Added a widget that displays a user's navigation level.

Changed the table classes so that a query supplied via useQuery() would execute at rendering time instead of on that function call.

Removed the ClassScanner class.

Moved all module-specific functionality to the module object and created a ModulePage class that all pages are derived from.

Fixed a bug with FloatField where not entering a decimal causes an error to occur.

Rewrote the templating system to use included PHP files instead of html files.

Added form error and warning boxes that are templatable.

Moved template PHP files to behind the web root.

Created the public/ directory that acts as the web root.

Switched query string building loops to use http_build_query();

Changed html minification to use Tidy.

Restructured the database to be cross-database compatible.

Added support for MySQL and SQLite 3.

Added password masking in the installation script and added a verification prompt for all passwords.

And last, but not least, numerous bugs related to all of these changes were fixed.