Changed incorrect form addError() calls to be addConfirmation() on successful submit.
Fixed a bug where plaintext elements were not being appended properly.

Added ability to specify form field descriptions.

Added the TableForm object.

Switched all admin form pages to use the TableForm object.

Added pre-baked class names to the textbox and password fields.

Added array format for a table cell provided to the Table object that allows a colspan and rowspan to be provided in addition to cell contents.

Expanded templating abilities of all table object.

Removed the form token from the admin login page.

Added functionality to the LinkList object to be able to mark an item as active.

Reskinned the admin module to the light style and set it as the default.

Restructured file structure of the resume admin pages to work better with namespaces.

Migrated remaining known html in strings generated in the controller to respective templates.

Added another EditTableForm object that acts as a separate, complimentary page that EditTable uses for add and edit operations.

Added a flag indicating of a field is interactive to the user for grouping and a function that returns only interactive fields.

Added a function to set required request variables from the general request object and enforce it when making request()->variable_name calls.

Changed the retrieval of request variables to look at $_POST first.

Add functions to allow urls of pages higher lower in the subdirectory level relative to the current page to be generated.

Made the action column of EditTable have 'Action' as its header.

And many other minor bugs that may have come up during development of these changes.