Added the ajax mode to handle ajax requests. This includes its own error object.
Modified the page mode and made it more flexible to accomodate the ajax mode.

Modified the .htaccess to rewrite ajax mode urls.

Added the Controller object to act as the base class for all controllers.

Made several properties of the framework static for easy retrieval from any part of the application.

Replaced all instances of the global framework() function with an equivalent static variable retrieval.

Removed the framework() global function.