Converted all page objects into controllers.
Moved the rendering of a page from the constructor to a setup() function in each controller.

Separated the building of a form in each controller that has one into the getForm() method.

Added ability to have public methods for use in framework modes that call a direct action such as ajax mode.

Added the submit() action method.

Brought the logic that loads the managed module if a module_id is present in the request into the constructor.

Added the css style and loading gif from the showLoading jQuery plugin.