Fixed a bug in the .htaccess that caused the page query string parameter to have the same value as subd.
Removed the retrieval of the module name from the url when converting the page url to a namespaced class path.

Fixed the Http object so it can automatically determine when to insert the specified module name into subd.

Added a static property in ModulePage to keep track of the module name that was loaded via the url.

Modified the admin module Home and Login controllers to specify module name overrides in their redirect urls.

Added caching of installed module names.