Fixed syntax errors in the installation script for SQLite.
Changed the goto triggered when an incorrect SQLite path has been entered to correctly point to the section prompting for the path.

Fixed a bug in PhoneNumber where empty value for a non-required field would still have its value validated against a phone number format.

Renamed the default style of the resume admin to 'light' to match the admin module.

Fixed the template path added in AdminPage to correctly replace 'admin' with '<module_name>/admin'.

Fixed an issue with Http::getInternalUrl where the provided module name was not added to the url when the current module is the default one.