Integrated EditTable into DataTable.
Implemented the EditTable JS object for AJAX interaction.

Updated all pages that utilize EditTable to work with the changes to the object.

Fixed a big with how the Action column was being added to a row and applied it to all EditTable pages.

Changed EditTableForm to work with the changes to EditTable.